fix: fix assertion error

+++ TGEngine_SOURCE/tgCapsuleCollider2D.cpp
@@ -47,18 +47,18 @@ namespace tg
         HPEN oldPen = (HPEN)SelectObject(hdc, greenPen);
 
         Arc(hdc
-            , pos.x + offset.x - radius, pos.y + offset.y - (height / 2)
-            , pos.x + offset.x + radius, pos.y + offset.y - (height / 2) + 2 * radius
-            , pos.x + offset.x + radius, pos.y + offset.y - (height / 2 - radius)
-            , pos.x + offset.x - radius, pos.y + offset.y - (height / 2 - radius));
+            , (int)(pos.x + offset.x - radius), (int)(pos.y + offset.y - (height / 2))
+            , (int)(pos.x + offset.x + radius), (int)(pos.y + offset.y - (height / 2) + 2 * radius)
+            , (int)(pos.x + offset.x + radius), (int)(pos.y + offset.y - (height / 2 - radius))
+            , (int)(pos.x + offset.x - radius), (int)(pos.y + offset.y - (height / 2 - radius)));
         ::Rectangle(hdc
-            , pos.x + offset.x - radius, pos.y + offset.y - (height / 2 - radius)
-            , pos.x + offset.x + radius, pos.y + offset.y + (height / 2 - radius));
+            , (int)(pos.x + offset.x - radius), (int)(pos.y + offset.y - (height / 2 - radius))
+            , (int)(pos.x + offset.x + radius), (int)(pos.y + offset.y + (height / 2 - radius)));
         Arc(hdc
-            , pos.x + offset.x - radius, pos.y + offset.y + (height / 2) - 2 * radius
-            , pos.x + offset.x + radius, pos.y + offset.y + (height / 2)
-            , pos.x + offset.x - radius, pos.y + offset.y + (height / 2 - radius)
-            , pos.x + offset.x + radius, pos.y + offset.y + (height / 2 - radius));
+            , (int)(pos.x + offset.x - radius), (int)(pos.y + offset.y + (height / 2) - 2 * radius)
+            , (int)(pos.x + offset.x + radius), (int)(pos.y + offset.y + (height / 2))
+            , (int)(pos.x + offset.x - radius), (int)(pos.y + offset.y + (height / 2 - radius))
+            , (int)(pos.x + offset.x + radius), (int)(pos.y + offset.y + (height / 2 - radius)));
 
         SelectObject(hdc, oldBrush);
         SelectObject(hdc, oldPen);
